+++ actix-files/src/lib.rs
@@ -223,17 +223,7 @@ mod tests {
 
     #[actix_rt::test]
     async fn test_named_file_non_ascii_file_name() {
-        let file = {
-            #[cfg(feature = "experimental-io-uring")]
-            {
-                crate::named::File::open("Cargo.toml").await.unwrap()
-            }
-
-            #[cfg(not(feature = "experimental-io-uring"))]
-            {
-                crate::named::File::open("Cargo.toml").unwrap()
-            }
-        };
+        let file = crate::named::File::open("Cargo.toml").unwrap();
 
         let mut file = NamedFile::from_file(file, "貨物.toml").unwrap();
         {